renames StaticSelectorMacros.h file to StaticSelectors.h
